feat: #436 - Add support for import and exports to NamespaceDeclaration
Create ModuledNode and share between SourceFile and NamespaceDeclaration
