+++ docs/details/exports.md
@@ -52,6 +52,79 @@ functionDeclaration.setIsExported(true); // be one
 functionDeclaration.setIsExported(false); // don't be one
 ```
 
+### Get default export symbol
+
+If it exists, the default export symbol can be retrieved from source file or module:
+
+```ts
+const defaultExportSymbol = sourceFile.getDefaultExportSymbol(); // returns: Symbol | undefined
+```
+
+### Remove default export
+
+Use:
+
+```ts
+sourceFile.removeDefaultExport();
+```
+
+Note: This is safe to call even when there is no default export.
+
+### Getting Exported Declarations
+
+The exported declarations of a file or module can be retrieved via `.getExportedDeclarations()`.
+
+For example, given the following setup:
+
+```ts
+// main.ts
+export * from "./classes";
+export {Interface1} from "./interfaces";
+
+class MainClass {}
+export default MainClass;
+
+// classes.ts
+export * from "./Class1";
+export * from "./Class2";
+
+// Class1.ts
+export class Class1 {}
+
+// Class2.ts
+export class Class2 {}
+
+// interfaces.ts
+export interface Interface1 {}
+export interface Interface2 {}
+```
+
+The following code:
+
+```ts
+import Project, {TypeGuards} from "ts-simple-ast";
+
+const project = new Project();
+project.addExistingSourceFiles("**/*.ts");
+const mainFile = project.getSourceFileOrThrow("main.ts");
+
+for (const declaration of mainFile.getExportedDeclarations()) {
+    if (TypeGuards.isClassDeclaration(declaration) || TypeGuards.isInterfaceDeclaration(declaration))
+        console.log(`Name: ${declaration.getName()}`);
+    else
+        throw new Error(`Not expected declaration kind: ${declaration.getKindName()}`);
+}
+```
+
+Outputs the following:
+
+```
+Name: MainClass
+Name: Class1
+Name: Class2
+Name: Interface1
+```
+
 ## Export Declarations
 
 Export declarations look like this:
