+++ src/codeBlockWriter/code-block-writer.ts
@@ -20,6 +20,8 @@ declare class CodeBlockWriter {
     private _stringCharStack;
     private _isInRegEx;
     private _isOnFirstLineOfBlock;
+    private static readonly _newLineRegEx;
+    private static readonly _spacesOrTabsRegEx;
     constructor(opts?: Partial<CodeBlockWriterOptions>);
     /**
      * Gets the options.
@@ -27,7 +29,7 @@ declare class CodeBlockWriter {
     getOptions(): CodeBlockWriterOptions;
     /**
      * Queues the indentation level for the next lines written.
-     * @param indentationLevel - Indentation level to be at.
+     * @param indentationLevel - Indentation level to queue.
      */
     queueIndentationLevel(indentationLevel: number): this;
     /**
@@ -79,9 +81,9 @@ declare class CodeBlockWriter {
     conditionalWriteLine(condition: boolean | undefined, text: string): this;
     /**
      * Writes a line of text.
-     * @param str - String to write.
+     * @param text - String to write.
      */
-    writeLine(str: string): this;
+    writeLine(text: string): this;
     /**
      * Writes a newline if the last line was not a newline.
      */
