+++ src/compiler/file/SourceFile.ts
@@ -1,25 +1,20 @@
 import * as errors from "../../errors";
 import { Directory } from "../../fileSystem";
 import { ProjectContext } from "../../ProjectContext";
-import { FormattingKind, getTextFromFormattingEdits, removeChildrenWithFormatting, replaceNodeText, replaceSourceFileForFilePathMove,
-    replaceSourceFileTextForFormatting } from "../../manipulation";
+import { getTextFromFormattingEdits, replaceNodeText, replaceSourceFileForFilePathMove, replaceSourceFileTextForFormatting } from "../../manipulation";
 import { getNextMatchingPos, getPreviousMatchingPos } from "../../manipulation/textSeek";
-import { ExportAssignmentStructure, ExportDeclarationStructure, ImportDeclarationStructure, SourceFileStructure, SourceFileSpecificStructure } from "../../structures";
+import { SourceFileStructure } from "../../structures";
 import { Constructor } from "../../types";
 import { LanguageVariant, ScriptTarget, SyntaxKind, ts } from "../../typescript";
-import { ArrayUtils, createHashSet, EventContainer, FileUtils, ModuleUtils, SourceFileReferenceContainer, StringUtils, TypeGuards } from "../../utils";
+import { ArrayUtils, EventContainer, FileUtils, ModuleUtils, SourceFileReferenceContainer, StringUtils, TypeGuards } from "../../utils";
 import { getBodyTextForStructure } from "../base/helpers";
-import { TextInsertableNode } from "../base";
+import { TextInsertableNode, ModuledNode } from "../base";
 import { callBaseSet } from "../callBaseSet";
 import { Node, Symbol } from "../common";
 import { StringLiteral } from "../literal";
 import { StatementedNode } from "../statement";
 import { Diagnostic, EmitOptionsBase, EmitOutput, EmitResult, FormatCodeSettings, UserPreferences, TextChange } from "../tools";
-import { ExportAssignment } from "./ExportAssignment";
-import { ExportDeclaration } from "./ExportDeclaration";
-import { ExportSpecifier } from "./ExportSpecifier";
 import { FileSystemRefreshResult } from "./FileSystemRefreshResult";
-import { ImportDeclaration } from "./ImportDeclaration";
 import { callBaseGetStructure } from "../callBaseGetStructure";
 
 export interface SourceFileCopyOptions {
@@ -42,8 +37,10 @@ export interface SourceFileReferences {
     referencingLiterals: StringLiteral[];
 }
 
-// todo: not sure why I need to explicitly type this in order to get VS to not complain... (TS 2.4.1)
-export const SourceFileBase: Constructor<StatementedNode> & Constructor<TextInsertableNode> & typeof Node = TextInsertableNode(StatementedNode(Node));
+// todo: not sure why I need to explicitly type this in order to get TS to not complain... (TS 2.4.1)
+export const SourceFileBase: Constructor<ModuledNode> & Constructor<StatementedNode> & Constructor<TextInsertableNode> & typeof Node = ModuledNode(
+    TextInsertableNode(StatementedNode(Node))
+);
 export class SourceFile extends SourceFileBase<ts.SourceFile> {
     /** @internal */
     private _isSaved = false;
@@ -71,25 +68,6 @@ export class SourceFile extends SourceFileBase<ts.SourceFile> {
         super(context, node, undefined as any);
         this.sourceFile = this;
         // end hack
-    }
-
-    /**
-     * Sets the node from a structure.
-     * @param structure - Structure to set the node with.
-     */
-    set(structure: Partial<SourceFileStructure>) {
-        callBaseSet(SourceFileBase.prototype, this, structure);
-
-        if (structure.imports != null) {
-            this.getImportDeclarations().forEach(d => d.remove());
-            this.addImportDeclarations(structure.imports);
-        }
-        if (structure.exports != null) {
-            this.getExportDeclarations().forEach(d => d.remove());
-            this.addExportDeclarations(structure.exports);
-        }
-
-        return this;
     }
 
     /**
@@ -511,358 +489,10 @@ export class SourceFile extends SourceFileBase<ts.SourceFile> {
     }
 
     /**
-     * Adds an import.
-     * @param structure - Structure that represents the import.
-     */
-    addImportDeclaration(structure: ImportDeclarationStructure) {
-        return this.addImportDeclarations([structure])[0];
-    }
-
-    /**
-     * Adds imports.
-     * @param structures - Structures that represent the imports.
-     */
-    addImportDeclarations(structures: ReadonlyArray<ImportDeclarationStructure>) {
-        const imports = this.getImportDeclarations();
-        const insertIndex = imports.length === 0 ? 0 : imports[imports.length - 1].getChildIndex() + 1;
-        return this.insertImportDeclarations(insertIndex, structures);
-    }
-
-    /**
-     * Insert an import.
-     * @param index - Child index to insert at.
-     * @param structure - Structure that represents the import.
-     */
-    insertImportDeclaration(index: number, structure: ImportDeclarationStructure) {
-        return this.insertImportDeclarations(index, [structure])[0];
-    }
-
-    /**
-     * Insert imports into a file.
-     * @param index - Child index to insert at.
-     * @param structures - Structures that represent the imports to insert.
-     */
-    insertImportDeclarations(index: number, structures: ReadonlyArray<ImportDeclarationStructure>): ImportDeclaration[] {
-        return this._insertChildren<ImportDeclaration, ImportDeclarationStructure>({
-            expectedKind: SyntaxKind.ImportDeclaration,
-            index,
-            structures,
-            write: (writer, info) => {
-                this._standardWrite(writer, info, () => {
-                    this.context.structurePrinterFactory.forImportDeclaration().printTexts(writer, structures);
-                }, {
-                    previousNewLine: previousMember => TypeGuards.isImportDeclaration(previousMember),
-                    nextNewLine: nextMember => TypeGuards.isImportDeclaration(nextMember)
-                });
-            }
-        });
-    }
-
-    /**
-     * Gets the first import declaration that matches a condition, or undefined if it doesn't exist.
-     * @param condition - Condition to get the import declaration by.
-     */
-    getImportDeclaration(condition: (importDeclaration: ImportDeclaration) => boolean): ImportDeclaration | undefined;
-    /**
-     * Gets the first import declaration that matches a module specifier, or undefined if it doesn't exist.
-     * @param module - Module specifier to get the import declaration by.
-     */
-    getImportDeclaration(moduleSpecifier: string): ImportDeclaration | undefined;
-    /** @internal */
-    getImportDeclaration(conditionOrModuleSpecifier: string | ((importDeclaration: ImportDeclaration) => boolean)): ImportDeclaration | undefined;
-    getImportDeclaration(conditionOrModuleSpecifier: string | ((importDeclaration: ImportDeclaration) => boolean)) {
-        return ArrayUtils.find(this.getImportDeclarations(), getCondition());
-
-        function getCondition() {
-            if (typeof conditionOrModuleSpecifier === "string")
-                return (dec: ImportDeclaration) => dec.getModuleSpecifierValue() === conditionOrModuleSpecifier;
-            else
-                return conditionOrModuleSpecifier;
-        }
-    }
-
-    /**
-     * Gets the first import declaration that matches a condition, or throws if it doesn't exist.
-     * @param condition - Condition to get the import declaration by.
-     */
-    getImportDeclarationOrThrow(condition: (importDeclaration: ImportDeclaration) => boolean): ImportDeclaration;
-    /**
-     * Gets the first import declaration that matches a module specifier, or throws if it doesn't exist.
-     * @param module - Module specifier to get the import declaration by.
-     */
-    getImportDeclarationOrThrow(moduleSpecifier: string): ImportDeclaration;
-    /** @internal */
-    getImportDeclarationOrThrow(conditionOrModuleSpecifier: string | ((importDeclaration: ImportDeclaration) => boolean)): ImportDeclaration;
-    getImportDeclarationOrThrow(conditionOrModuleSpecifier: string | ((importDeclaration: ImportDeclaration) => boolean)) {
-        return errors.throwIfNullOrUndefined(this.getImportDeclaration(conditionOrModuleSpecifier), "Expected to find an import with the provided condition.");
-    }
-
-    /**
-     * Get the file's import declarations.
-     */
-    getImportDeclarations(): ImportDeclaration[] {
-        // todo: remove type assertion
-        return this.getChildSyntaxListOrThrow().getChildrenOfKind(SyntaxKind.ImportDeclaration);
-    }
-
-    /**
-     * Add export declarations.
-     * @param structure - Structure that represents the export.
-     */
-    addExportDeclaration(structure: ExportDeclarationStructure) {
-        return this.addExportDeclarations([structure])[0];
-    }
-
-    /**
-     * Add export declarations.
-     * @param structures - Structures that represent the exports.
-     */
-    addExportDeclarations(structures: ReadonlyArray<ExportDeclarationStructure>) {
-        // always insert at end of file because of export {Identifier}; statements
-        return this.insertExportDeclarations(this.getChildSyntaxListOrThrow().getChildCount(), structures);
-    }
-
-    /**
-     * Insert an export declaration.
-     * @param index - Child index to insert at.
-     * @param structure - Structure that represents the export.
-     */
-    insertExportDeclaration(index: number, structure: ExportDeclarationStructure) {
-        return this.insertExportDeclarations(index, [structure])[0];
-    }
-
-    /**
-     * Insert export declarations into a file.
-     * @param index - Child index to insert at.
-     * @param structures - Structures that represent the exports to insert.
-     */
-    insertExportDeclarations(index: number, structures: ReadonlyArray<ExportDeclarationStructure>): ExportDeclaration[] {
-        return this._insertChildren<ExportDeclaration, ExportDeclarationStructure>({
-            expectedKind: SyntaxKind.ExportDeclaration,
-            index,
-            structures,
-            write: (writer, info) => {
-                this._standardWrite(writer, info, () => {
-                    this.context.structurePrinterFactory.forExportDeclaration().printTexts(writer, structures);
-                }, {
-                    previousNewLine: previousMember => TypeGuards.isExportDeclaration(previousMember),
-                    nextNewLine: nextMember => TypeGuards.isExportDeclaration(nextMember)
-                });
-            }
-        });
-    }
-
-    /**
-     * Gets the first export declaration that matches a condition, or undefined if it doesn't exist.
-     * @param condition - Condition to get the export declaration by.
-     */
-    getExportDeclaration(condition: (exportDeclaration: ExportDeclaration) => boolean): ExportDeclaration | undefined;
-    /**
-     * Gets the first export declaration that matches a module specifier, or undefined if it doesn't exist.
-     * @param module - Module specifier to get the export declaration by.
-     */
-    getExportDeclaration(moduleSpecifier: string): ExportDeclaration | undefined;
-    /** @internal */
-    getExportDeclaration(conditionOrModuleSpecifier: string | ((exportDeclaration: ExportDeclaration) => boolean)): ExportDeclaration | undefined;
-    getExportDeclaration(conditionOrModuleSpecifier: string | ((exportDeclaration: ExportDeclaration) => boolean)) {
-        return ArrayUtils.find(this.getExportDeclarations(), getCondition());
-
-        function getCondition() {
-            if (typeof conditionOrModuleSpecifier === "string")
-                return (dec: ExportDeclaration) => dec.getModuleSpecifierValue() === conditionOrModuleSpecifier;
-            else
-                return conditionOrModuleSpecifier;
-        }
-    }
-
-    /**
-     * Gets the first export declaration that matches a condition, or throws if it doesn't exist.
-     * @param condition - Condition to get the export declaration by.
-     */
-    getExportDeclarationOrThrow(condition: (exportDeclaration: ExportDeclaration) => boolean): ExportDeclaration;
-    /**
-     * Gets the first export declaration that matches a module specifier, or throws if it doesn't exist.
-     * @param module - Module specifier to get the export declaration by.
-     */
-    getExportDeclarationOrThrow(moduleSpecifier: string): ExportDeclaration;
-    /** @internal */
-    getExportDeclarationOrThrow(conditionOrModuleSpecifier: string | ((exportDeclaration: ExportDeclaration) => boolean)): ExportDeclaration;
-    getExportDeclarationOrThrow(conditionOrModuleSpecifier: string | ((exportDeclaration: ExportDeclaration) => boolean)) {
-        return errors.throwIfNullOrUndefined(this.getExportDeclaration(conditionOrModuleSpecifier), "Expected to find an export declaration with the provided condition.");
-    }
-
-    /**
-     * Get the file's export declarations.
-     */
-    getExportDeclarations(): ExportDeclaration[] {
-        return this.getChildSyntaxListOrThrow().getChildrenOfKind(SyntaxKind.ExportDeclaration);
-    }
-
-    /**
-     * Gets the export symbols of the source file.
-     */
-    getExportSymbols(): Symbol[] {
-        const symbol = this.getSymbol();
-        return symbol == null ? [] : this.context.typeChecker.getExportsOfModule(symbol);
-    }
-
-    /**
-     * Gets all the declarations that are exported from the file.
-     *
-     * This will include declarations that are transitively exported from other files. If you mean to get the export
-     * declarations then use sourceFile.getExportDeclarations().
-     */
-    getExportedDeclarations(): Node[] {
-        const exportSymbols = this.getExportSymbols();
-        return ArrayUtils.from(getDeclarationsForSymbols());
-
-        function* getDeclarationsForSymbols() {
-            const handledDeclarations = createHashSet<Node>();
-
-            for (const symbol of exportSymbols)
-                for (const declaration of symbol.getDeclarations())
-                    yield* getDeclarationHandlingExportSpecifiers(declaration);
-
-            function* getDeclarationHandlingExportSpecifiers(declaration: Node): IterableIterator<Node> {
-                if (handledDeclarations.has(declaration))
-                    return;
-                handledDeclarations.add(declaration);
-
-                if (declaration.getKind() === SyntaxKind.ExportSpecifier) {
-                    for (const d of (declaration as ExportSpecifier).getLocalTargetDeclarations())
-                        yield* getDeclarationHandlingExportSpecifiers(d);
-                }
-                else if (declaration.getKind() === SyntaxKind.ExportAssignment) {
-                    const identifier = (declaration as ExportAssignment).getExpression();
-                    if (identifier == null || identifier.getKind() !== SyntaxKind.Identifier)
-                        return;
-                    const symbol = identifier.getSymbol();
-                    if (symbol == null)
-                        return;
-                    for (const d of symbol.getDeclarations())
-                        yield* getDeclarationHandlingExportSpecifiers(d);
-                }
-                else
-                    yield declaration;
-            }
-        }
-    }
-
-    /**
-     * Add export assignments.
-     * @param structure - Structure that represents the export.
-     */
-    addExportAssignment(structure: ExportAssignmentStructure) {
-        return this.addExportAssignments([structure])[0];
-    }
-
-    /**
-     * Add export assignments.
-     * @param structures - Structures that represent the exports.
-     */
-    addExportAssignments(structures: ReadonlyArray<ExportAssignmentStructure>) {
-        // always insert at end of file because of export {Identifier}; statements
-        return this.insertExportAssignments(this.getChildSyntaxListOrThrow().getChildCount(), structures);
-    }
-
-    /**
-     * Insert an export assignment.
-     * @param index - Child index to insert at.
-     * @param structure - Structure that represents the export.
-     */
-    insertExportAssignment(index: number, structure: ExportAssignmentStructure) {
-        return this.insertExportAssignments(index, [structure])[0];
-    }
-
-    /**
-     * Insert export assignments into a file.
-     * @param index - Child index to insert at.
-     * @param structures - Structures that represent the exports to insert.
-     */
-    insertExportAssignments(index: number, structures: ReadonlyArray<ExportAssignmentStructure>): ExportAssignment[] {
-        return this._insertChildren<ExportAssignment, ExportAssignmentStructure>({
-            expectedKind: SyntaxKind.ExportAssignment,
-            index,
-            structures,
-            write: (writer, info) => {
-                this._standardWrite(writer, info, () => {
-                    this.context.structurePrinterFactory.forExportAssignment().printTexts(writer, structures);
-                }, {
-                    previousNewLine: previousMember => TypeGuards.isExportAssignment(previousMember),
-                    nextNewLine: nextMember => TypeGuards.isExportAssignment(nextMember)
-                });
-            }
-        });
-    }
-
-    /**
-     * Gets the first export assignment that matches a condition, or undefined if it doesn't exist.
-     * @param condition - Condition to get the export assignment by.
-     */
-    getExportAssignment(condition: (exportAssignment: ExportAssignment) => boolean): ExportAssignment | undefined {
-        return ArrayUtils.find(this.getExportAssignments(), condition);
-    }
-
-    /**
-     * Gets the first export assignment that matches a condition, or throws if it doesn't exist.
-     * @param condition - Condition to get the export assignment by.
-     */
-    getExportAssignmentOrThrow(condition: (exportAssignment: ExportAssignment) => boolean): ExportAssignment {
-        return errors.throwIfNullOrUndefined(this.getExportAssignment(condition), "Expected to find an export assignment with the provided condition.");
-    }
-
-    /**
-     * Get the file's export assignments.
-     */
-    getExportAssignments(): ExportAssignment[] {
-        return this.getChildSyntaxListOrThrow().getChildrenOfKind(SyntaxKind.ExportAssignment);
-    }
-
-    /**
-     * Gets the default export symbol of the file.
-     */
-    getDefaultExportSymbol(): Symbol | undefined {
-        const sourceFileSymbol = this.getSymbol();
-
-        // will be undefined when the source file doesn't have an export
-        if (sourceFileSymbol == null)
-            return undefined;
-
-        return sourceFileSymbol.getExportByName("default");
-    }
-
-    /**
-     * Gets the default export symbol of the file or throws if it doesn't exist.
-     */
-    getDefaultExportSymbolOrThrow(): Symbol {
-        return errors.throwIfNullOrUndefined(this.getDefaultExportSymbol(), "Expected to find a default export symbol");
-    }
-
-    /**
      * Gets the pre-emit diagnostics of the specified source file.
      */
     getPreEmitDiagnostics(): Diagnostic[] {
         return this.context.getPreEmitDiagnostics(this);
-    }
-
-    /**
-     * Removes any "export default";
-     */
-    removeDefaultExport(defaultExportSymbol?: Symbol | undefined): this {
-        defaultExportSymbol = defaultExportSymbol || this.getDefaultExportSymbol();
-
-        if (defaultExportSymbol == null)
-            return this;
-
-        const declaration = defaultExportSymbol.getDeclarations()[0];
-        if (declaration.compilerNode.kind === SyntaxKind.ExportAssignment)
-            removeChildrenWithFormatting({ children: [declaration], getSiblingFormatting: () => FormattingKind.Newline });
-        else if (TypeGuards.isModifierableNode(declaration)) {
-            declaration.toggleModifier("default", false);
-            declaration.toggleModifier("export", false);
-        }
-
-        return this;
     }
 
     /**
@@ -1057,6 +687,16 @@ export class SourceFile extends SourceFileBase<ts.SourceFile> {
             replacingLength: this.getFullWidth(),
             newText: getTextFromFormattingEdits(this, textChanges)
         });
+        return this;
+    }
+
+    /**
+     * Sets the node from a structure.
+     * @param structure - Structure to set the node with.
+     */
+    set(structure: Partial<SourceFileStructure>) {
+        callBaseSet(SourceFileBase.prototype, this, structure);
+
         return this;
     }
