+++ src/structures/base/ModuledNodeStructure.ts
@@ -1,0 +1,6 @@
+﻿import { ImportDeclarationStructure, ExportDeclarationStructure } from "../file";
+
+export interface ModuledNodeStructure {
+    imports?: ImportDeclarationStructure[];
+    exports?: ExportDeclarationStructure[];
+}
