+++ src/structures/namespace/NamespaceDeclarationStructure.ts
@@ -1,9 +1,10 @@
-﻿import { AmbientableNodeStructure, BodiedNodeStructure, ExportableNodeStructure, JSDocableNodeStructure, NamedNodeStructure } from "../base";
+﻿import { AmbientableNodeStructure, BodiedNodeStructure, ExportableNodeStructure, JSDocableNodeStructure, NamedNodeStructure,
+    ModuledNodeStructure } from "../base";
 import { StatementedNodeStructure } from "../statement";
 
 export interface NamespaceDeclarationStructure
     extends NamedNodeStructure, NamespaceDeclarationSpecificStructure, JSDocableNodeStructure, AmbientableNodeStructure, ExportableNodeStructure, StatementedNodeStructure,
-        BodiedNodeStructure
+        BodiedNodeStructure, ModuledNodeStructure
 {
 }
