+++ src/tests/compiler/file/sourceFileTests.ts
@@ -4,7 +4,7 @@ import { Node, EmitResult, ExportAssignment, ExportDeclaration, FileSystemRefres
 import * as errors from "../../../errors";
 import { IndentationText, ManipulationSettings } from "../../../options";
 import { Project } from "../../../Project";
-import { ExportAssignmentStructure, ExportDeclarationStructure, ImportDeclarationStructure, SourceFileSpecificStructure } from "../../../structures";
+import { ExportAssignmentStructure, ExportDeclarationStructure, ImportDeclarationStructure, SourceFileStructure } from "../../../structures";
 import { CompilerOptions, LanguageVariant, ModuleResolutionKind, NewLineKind, ScriptTarget } from "../../../typescript";
 import { getFileSystemHostWithFiles } from "../../testHelpers";
 import { getInfoFromText } from "../testHelpers";
@@ -473,483 +473,6 @@ describe(nameof(SourceFile), () => {
         });
     });
 
-    describe(nameof<SourceFile>(n => n.insertImportDeclarations), () => {
-        function doTest(startCode: string, index: number, structures: ImportDeclarationStructure[], expectedCode: string, useSingleQuotes = false) {
-            const { sourceFile, project } = getInfoFromText(startCode);
-            if (useSingleQuotes)
-                project.manipulationSettings.set({ quoteKind: QuoteKind.Single });
-            const result = sourceFile.insertImportDeclarations(index, structures);
-            expect(result.length).to.equal(structures.length);
-            expect(sourceFile.getFullText()).to.equal(expectedCode);
-        }
-
-        it("should insert the different kinds of imports", () => {
-            doTest("", 0, [
-                { moduleSpecifier: "./test" },
-                { defaultImport: "identifier", moduleSpecifier: "./test" },
-                { defaultImport: "identifier", namespaceImport: "name", moduleSpecifier: "./test" },
-                { defaultImport: "identifier", namedImports: ["name1", { name: "name" }, { name: "name", alias: "alias" }], moduleSpecifier: "./test" },
-                { namedImports: ["name"], moduleSpecifier: "./test" },
-                { namespaceImport: "name", moduleSpecifier: "./test" }
-            ], [
-                `import "./test";`,
-                `import identifier from "./test";`,
-                `import identifier, * as name from "./test";`,
-                `import identifier, { name1, name, name as alias } from "./test";`,
-                `import { name } from "./test";`,
-                `import * as name from "./test";`
-            ].join("\n") + "\n");
-        });
-
-        it("should throw when specifying a namespace import and named imports", () => {
-            const { sourceFile } = getInfoFromText("");
-
-            expect(() => {
-                sourceFile.insertImportDeclarations(0, [{ namespaceImport: "name", namedImports: ["name"], moduleSpecifier: "file" }]);
-            }).to.throw();
-        });
-
-        it("should insert an import if the file was read with a utf-8 bom", () => {
-            doTest("\uFEFF", 0, [{ moduleSpecifier: "./test" }], `import "./test";\n`);
-        });
-
-        it("should insert at the beginning and use single quotes when specified", () => {
-            doTest(`export class Class {}\n`, 0, [{ moduleSpecifier: "./test" }], `import './test';\n\nexport class Class {}\n`, true);
-        });
-
-        it("should insert in the middle", () => {
-            doTest(`import "./file1";\nimport "./file3";\n`, 1, [{ moduleSpecifier: "./file2" }], `import "./file1";\nimport "./file2";\nimport "./file3";\n`);
-        });
-
-        it("should insert at the end", () => {
-            doTest(`export class Class {}\n`, 1, [{ moduleSpecifier: "./test" }], `export class Class {}\n\nimport "./test";\n`);
-        });
-    });
-
-    describe(nameof<SourceFile>(n => n.insertImportDeclaration), () => {
-        function doTest(startCode: string, index: number, structure: ImportDeclarationStructure, expectedCode: string) {
-            const { sourceFile } = getInfoFromText(startCode);
-            const result = sourceFile.insertImportDeclaration(index, structure);
-            expect(result).to.be.instanceOf(ImportDeclaration);
-            expect(sourceFile.getText()).to.equal(expectedCode);
-        }
-
-        it("should insert at the specified position", () => {
-            doTest(`import "./file1";\nimport "./file3";\n`, 1, { moduleSpecifier: "./file2" }, `import "./file1";\nimport "./file2";\nimport "./file3";\n`);
-        });
-    });
-
-    describe(nameof<SourceFile>(n => n.addImportDeclaration), () => {
-        function doTest(startCode: string, structure: ImportDeclarationStructure, expectedCode: string) {
-            const { sourceFile } = getInfoFromText(startCode);
-            const result = sourceFile.addImportDeclaration(structure);
-            expect(result).to.be.instanceOf(ImportDeclaration);
-            expect(sourceFile.getText()).to.equal(expectedCode);
-        }
-
-        it("should add at the last import if one exists", () => {
-            doTest(`import "./file1";\nimport "./file2";\n\nexport class MyClass {}\n`, { moduleSpecifier: "./file3" },
-                `import "./file1";\nimport "./file2";\nimport "./file3";\n\nexport class MyClass {}\n`);
-        });
-
-        it("should add at the start if no imports exists", () => {
-            doTest(`export class MyClass {}\n`, { moduleSpecifier: "./file" },
-                `import "./file";\n\nexport class MyClass {}\n`);
-        });
-    });
-
-    describe(nameof<SourceFile>(n => n.addImportDeclarations), () => {
-        function doTest(startCode: string, structures: ImportDeclarationStructure[], expectedCode: string) {
-            const { sourceFile } = getInfoFromText(startCode);
-            const result = sourceFile.addImportDeclarations(structures);
-            expect(result.length).to.equal(structures.length);
-            expect(sourceFile.getText()).to.equal(expectedCode);
-        }
-
-        it("should add at the last import if one exists", () => {
-            doTest(`import "./file1";\n\nexport class MyClass {}\n`, [{ moduleSpecifier: "./file2" }, { moduleSpecifier: "./file3" }],
-                `import "./file1";\nimport "./file2";\nimport "./file3";\n\nexport class MyClass {}\n`);
-        });
-
-        it("should add at the start if no imports exists", () => {
-            doTest(`export class MyClass {}\n`, [{ moduleSpecifier: "./file1" }, { moduleSpecifier: "./file2" }],
-                `import "./file1";\nimport "./file2";\n\nexport class MyClass {}\n`);
-        });
-    });
-
-    describe(nameof<SourceFile>(n => n.getImportDeclarations), () => {
-        it("should get the import declarations", () => {
-            const { sourceFile } = getInfoFromText("import myImport from 'test'; import {next} from './test';");
-            expect(sourceFile.getImportDeclarations().length).to.equal(2);
-            expect(sourceFile.getImportDeclarations()[0]).to.be.instanceOf(ImportDeclaration);
-        });
-    });
-
-    describe(nameof<SourceFile>(n => n.getImportDeclaration), () => {
-        function doTest(text: string, conditionOrModuleSpecifier: string | ((importDeclaration: ImportDeclaration) => boolean), expected: string | undefined) {
-            const { sourceFile } = getInfoFromText(text);
-            const result = sourceFile.getImportDeclaration(conditionOrModuleSpecifier);
-            if (expected == null)
-                expect(result).to.be.undefined;
-            else
-                expect(result!.getText()).to.equal(expected);
-        }
-
-        it("should get the import declaration", () => {
-            doTest("import myImport from 'test'; import {next} from './test';", i => i.getDefaultImport() != null, "import myImport from 'test';");
-        });
-
-        it("should get the import declaration when providing module specifier text", () => {
-            doTest("import myImport from 'test'; import {next} from './test';", "./test", "import {next} from './test';");
-        });
-
-        it("should return undefined when not exists", () => {
-            doTest("import myImport from 'test';", "asdfasdf", undefined);
-        });
-    });
-
-    describe(nameof<SourceFile>(n => n.getImportDeclarationOrThrow), () => {
-        function doTest(text: string, conditionOrModuleSpecifier: string | ((importDeclaration: ImportDeclaration) => boolean), expected: string | undefined) {
-            const { sourceFile } = getInfoFromText(text);
-            if (expected == null)
-                expect(() => sourceFile.getImportDeclarationOrThrow(conditionOrModuleSpecifier)).to.throw();
-            else
-                expect(sourceFile.getImportDeclarationOrThrow(conditionOrModuleSpecifier).getText()).to.equal(expected);
-        }
-
-        it("should get the import declaration", () => {
-            doTest("import myImport from 'test'; import {next} from './test';", i => i.getDefaultImport() != null, "import myImport from 'test';");
-        });
-
-        it("should get the import declaration when providing module specifier text", () => {
-            doTest("import myImport from 'test'; import {next} from './test';", "./test", "import {next} from './test';");
-        });
-
-        it("should throw when not exists", () => {
-            doTest("import myImport from 'test';", "asdfasdf", undefined);
-        });
-    });
-
-    describe(nameof<SourceFile>(n => n.insertExportDeclarations), () => {
-        function doTest(startCode: string, index: number, structures: ExportDeclarationStructure[], expectedCode: string) {
-            const { sourceFile } = getInfoFromText(startCode);
-            const result = sourceFile.insertExportDeclarations(index, structures);
-            expect(result.length).to.equal(structures.length);
-            expect(sourceFile.getText()).to.equal(expectedCode);
-        }
-
-        it("should insert the different kinds of exports", () => {
-            doTest("", 0, [
-                { moduleSpecifier: "./test" },
-                { namedExports: ["name1", { name: "name" }, { name: "name", alias: "alias" }], moduleSpecifier: "./test" },
-                { namedExports: ["name"] },
-                {}
-            ], [
-                `export * from "./test";`,
-                `export { name1, name, name as alias } from "./test";`,
-                `export { name };`,
-                `export { };`
-            ].join("\n") + "\n");
-        });
-
-        it("should insert at the beginning", () => {
-            doTest(`export class Class {}\n`, 0, [{ moduleSpecifier: "./test" }], `export * from "./test";\n\nexport class Class {}\n`);
-        });
-
-        it("should insert in the middle", () => {
-            doTest(`export * from "./file1";\nexport * from "./file3";\n`, 1, [{ moduleSpecifier: "./file2" }],
-                `export * from "./file1";\nexport * from "./file2";\nexport * from "./file3";\n`);
-        });
-
-        it("should insert at the end", () => {
-            doTest(`export class Class {}\n`, 1, [{ moduleSpecifier: "./test" }], `export class Class {}\n\nexport * from "./test";\n`);
-        });
-    });
-
-    describe(nameof<SourceFile>(n => n.insertExportDeclaration), () => {
-        function doTest(startCode: string, index: number, structure: ExportDeclarationStructure, expectedCode: string) {
-            const { sourceFile } = getInfoFromText(startCode);
-            const result = sourceFile.insertExportDeclaration(index, structure);
-            expect(result).to.be.instanceOf(ExportDeclaration);
-            expect(sourceFile.getText()).to.equal(expectedCode);
-        }
-
-        it("should insert at the specified position", () => {
-            doTest(`export * from "./file1";\nexport * from "./file3";\n`, 1, { moduleSpecifier: "./file2" },
-                `export * from "./file1";\nexport * from "./file2";\nexport * from "./file3";\n`);
-        });
-    });
-
-    describe(nameof<SourceFile>(n => n.addExportDeclaration), () => {
-        function doTest(startCode: string, structure: ExportDeclarationStructure, expectedCode: string) {
-            const { sourceFile } = getInfoFromText(startCode);
-            const result = sourceFile.addExportDeclaration(structure);
-            expect(result).to.be.instanceOf(ExportDeclaration);
-            expect(sourceFile.getText()).to.equal(expectedCode);
-        }
-
-        it("should always add at the end of the file", () => {
-            doTest(`export class MyClass {}\n`, { moduleSpecifier: "./file" },
-                `export class MyClass {}\n\nexport * from "./file";\n`);
-        });
-    });
-
-    describe(nameof<SourceFile>(n => n.addExportDeclarations), () => {
-        function doTest(startCode: string, structures: ExportDeclarationStructure[], expectedCode: string) {
-            const { sourceFile } = getInfoFromText(startCode);
-            const result = sourceFile.addExportDeclarations(structures);
-            expect(result.length).to.equal(structures.length);
-            expect(sourceFile.getText()).to.equal(expectedCode);
-        }
-
-        it("should add multiple", () => {
-            doTest(`export class MyClass {}\n`, [{ moduleSpecifier: "./file1" }, { moduleSpecifier: "./file2" }],
-                `export class MyClass {}\n\nexport * from "./file1";\nexport * from "./file2";\n`);
-        });
-    });
-
-    describe(nameof<SourceFile>(n => n.getExportDeclarations), () => {
-        it("should get the export declarations", () => {
-            const { sourceFile } = getInfoFromText("export * from 'test'; export {next} from './test';");
-            expect(sourceFile.getExportDeclarations().length).to.equal(2);
-            expect(sourceFile.getExportDeclarations()[0]).to.be.instanceOf(ExportDeclaration);
-        });
-    });
-
-    describe(nameof<SourceFile>(n => n.getExportDeclaration), () => {
-        function doTest(text: string, conditionOrModuleSpecifier: string | ((importDeclaration: ExportDeclaration) => boolean), expected: string | undefined) {
-            const { sourceFile } = getInfoFromText(text);
-            const result = sourceFile.getExportDeclaration(conditionOrModuleSpecifier);
-            if (expected == null)
-                expect(result).to.be.undefined;
-            else
-                expect(result!.getText()).to.equal(expected);
-        }
-
-        it("should get when exists", () => {
-            doTest("export * from 'test'; export {next} from './test';", e => e.isNamespaceExport(), "export * from 'test';");
-        });
-
-        it("should get when exists by module specifier", () => {
-            doTest("export * from 'test'; export {next} from './test';", "./test", "export {next} from './test';");
-        });
-
-        it("should return undefined when not exists", () => {
-            doTest("export * from 'test'; export {next} from './test';", "not-exists", undefined);
-        });
-    });
-
-    describe(nameof<SourceFile>(n => n.getExportDeclarationOrThrow), () => {
-        function doTest(text: string, conditionOrModuleSpecifier: string | ((importDeclaration: ExportDeclaration) => boolean), expected: string | undefined) {
-            const { sourceFile } = getInfoFromText(text);
-            if (expected == null)
-                expect(() => sourceFile.getExportDeclarationOrThrow(conditionOrModuleSpecifier)).to.throw();
-            else
-                expect(sourceFile.getExportDeclarationOrThrow(conditionOrModuleSpecifier).getText()).to.equal(expected);
-        }
-
-        it("should get when exists", () => {
-            doTest("export * from 'test'; export {next} from './test';", e => e.isNamespaceExport(), "export * from 'test';");
-        });
-
-        it("should get when exists by module specifier", () => {
-            doTest("export * from 'test'; export {next} from './test';", "./test", "export {next} from './test';");
-        });
-
-        it("should throw when not exists", () => {
-            doTest("export * from 'test'; export {next} from './test';", "not-exists", undefined);
-        });
-    });
-
-    describe(nameof<SourceFile>(n => n.getExportedDeclarations), () => {
-        it("should get the exported declarations", () => {
-            const project = new Project({ useVirtualFileSystem: true });
-            const mainSourceFile = project.createSourceFile("main.ts", `export * from "./class";\nexport {OtherClass} from "./otherClass";\nexport * from "./barrel";\n` +
-                "export class MainFileClass {}\nexport default MainFileClass;");
-            project.createSourceFile("class.ts", `export class Class {} export class MyClass {}`);
-            project.createSourceFile("otherClass.ts", `export class OtherClass {}\nexport class InnerClass {}`);
-            project.createSourceFile("barrel.ts", `export * from "./subBarrel";`);
-            project.createSourceFile("subBarrel.ts", `export * from "./subFile";\nexport {SubClass2 as Test} from "./subFile2";\n` +
-                `export {default as SubClass3} from "./subFile3"`);
-            project.createSourceFile("subFile.ts", `export class SubClass {}`);
-            project.createSourceFile("subFile2.ts", `export class SubClass2 {}`);
-            project.createSourceFile("subFile3.ts", `class SubClass3 {}\nexport default SubClass3;`);
-
-            expect(mainSourceFile.getExportedDeclarations().map(d => (d as any).getName()).sort())
-                .to.deep.equal(["MainFileClass", "OtherClass", "Class", "MyClass", "SubClass", "SubClass2", "SubClass3"].sort());
-        });
-
-        function doTest(text: string, expectedDeclarationNames: string[]) {
-            const project = new Project({ useVirtualFileSystem: true });
-            const mainSourceFile = project.createSourceFile("main.ts", text);
-
-            expect(mainSourceFile.getExportedDeclarations().map(d => (d as any).getName()).sort())
-                .to.deep.equal(expectedDeclarationNames.sort());
-        }
-
-        it("should get the exported declaration when there's only a default export using an export assignment", () => {
-            doTest("class MainFileClass {}\nexport default MainFileClass;", ["MainFileClass"]);
-        });
-
-        it("should not error for an empty file", () => {
-            doTest("", []);
-        });
-    });
-
-    describe(nameof<SourceFile>(n => n.insertExportAssignments), () => {
-        function doTest(startCode: string, index: number, structures: ExportAssignmentStructure[], expectedCode: string) {
-            const { sourceFile } = getInfoFromText(startCode);
-            const result = sourceFile.insertExportAssignments(index, structures);
-            expect(result.length).to.equal(structures.length);
-            expect(sourceFile.getText()).to.equal(expectedCode);
-        }
-
-        it("should insert the different kinds of exports", () => {
-            doTest("", 0, [
-                { expression: "5" },
-                { isExportEquals: true, expression: writer => writer.write("6") },
-                { isExportEquals: false, expression: "name" }
-            ], [
-                `export = 5;`,
-                `export = 6;`,
-                `export default name;`
-            ].join("\n") + "\n");
-        });
-
-        it("should insert at the beginning", () => {
-            doTest(`export class Class {}\n`, 0, [{ expression: "5" }], `export = 5;\n\nexport class Class {}\n`);
-        });
-
-        it("should insert in the middle", () => {
-            doTest(`export * from "./file1";\nexport = 6;\n`, 1, [{ expression: "5" }],
-                `export * from "./file1";\n\nexport = 5;\nexport = 6;\n`);
-        });
-
-        it("should insert at the end", () => {
-            doTest(`export class Class {}\n`, 1, [{ expression: "5" }], `export class Class {}\n\nexport = 5;\n`);
-        });
-    });
-
-    describe(nameof<SourceFile>(n => n.insertExportAssignment), () => {
-        function doTest(startCode: string, index: number, structure: ExportAssignmentStructure, expectedCode: string) {
-            const { sourceFile } = getInfoFromText(startCode);
-            const result = sourceFile.insertExportAssignment(index, structure);
-            expect(result).to.be.instanceOf(ExportAssignment);
-            expect(sourceFile.getText()).to.equal(expectedCode);
-        }
-
-        it("should insert at the specified position", () => {
-            doTest(`export * from "./file1";\nexport = 6;\n`, 1, { expression: "5" },
-                `export * from "./file1";\n\nexport = 5;\nexport = 6;\n`);
-        });
-    });
-
-    describe(nameof<SourceFile>(n => n.addExportAssignment), () => {
-        function doTest(startCode: string, structure: ExportAssignmentStructure, expectedCode: string) {
-            const { sourceFile } = getInfoFromText(startCode);
-            const result = sourceFile.addExportAssignment(structure);
-            expect(result).to.be.instanceOf(ExportAssignment);
-            expect(sourceFile.getText()).to.equal(expectedCode);
-        }
-
-        it("should always add at the end of the file", () => {
-            doTest(`export class MyClass {}\n`, { expression: "5" },
-                `export class MyClass {}\n\nexport = 5;\n`);
-        });
-    });
-
-    describe(nameof<SourceFile>(n => n.addExportAssignments), () => {
-        function doTest(startCode: string, structures: ExportAssignmentStructure[], expectedCode: string) {
-            const { sourceFile } = getInfoFromText(startCode);
-            const result = sourceFile.addExportAssignments(structures);
-            expect(result.length).to.equal(structures.length);
-            expect(sourceFile.getText()).to.equal(expectedCode);
-        }
-
-        it("should add multiple", () => {
-            doTest(`export class MyClass {}\n`, [{ expression: "5" }, { expression: "6" }],
-                `export class MyClass {}\n\nexport = 5;\nexport = 6;\n`);
-        });
-    });
-
-    describe(nameof<SourceFile>(n => n.getExportAssignments), () => {
-        it("should get the export declarations", () => {
-            const { sourceFile } = getInfoFromText("export = 5; export = 6;");
-            expect(sourceFile.getExportAssignments().length).to.equal(2);
-            expect(sourceFile.getExportAssignments()[0]).to.be.instanceOf(ExportAssignment);
-        });
-    });
-
-    describe(nameof<SourceFile>(n => n.getExportAssignment), () => {
-        it("should get the export declaration", () => {
-            const { sourceFile } = getInfoFromText("export = 5; export default 6;");
-            expect(sourceFile.getExportAssignment(e => !e.isExportEquals())!.getText()).to.equal("export default 6;");
-        });
-    });
-
-    describe(nameof<SourceFile>(n => n.getExportAssignmentOrThrow), () => {
-        it("should get the export declaration", () => {
-            const { sourceFile } = getInfoFromText("export = 5; export default 6;");
-            expect(sourceFile.getExportAssignmentOrThrow(e => !e.isExportEquals()).getText()).to.equal("export default 6;");
-        });
-
-        it("should throw when not exists", () => {
-            const { sourceFile } = getInfoFromText("");
-            expect(() => sourceFile.getExportAssignmentOrThrow(e => false)).to.throw();
-        });
-    });
-
-    describe(nameof<SourceFile>(n => n.getDefaultExportSymbol), () => {
-        it("should return undefined when there's no default export", () => {
-            const { sourceFile } = getInfoFromText("");
-            expect(sourceFile.getDefaultExportSymbol()).to.be.undefined;
-        });
-
-        it("should return the default export symbol when one exists", () => {
-            const { sourceFile } = getInfoFromText("export default class Identifier {}");
-            const defaultExportSymbol = sourceFile.getDefaultExportSymbol()!;
-            expect(defaultExportSymbol.getName()).to.equal("default");
-        });
-
-        it("should return the default export symbol when default exported on a separate statement", () => {
-            const { sourceFile } = getInfoFromText("class Identifier {}\nexport default Identifier;");
-            const defaultExportSymbol = sourceFile.getDefaultExportSymbol()!;
-            expect(defaultExportSymbol.getName()).to.equal("default");
-        });
-    });
-
-    describe(nameof<SourceFile>(n => n.getDefaultExportSymbolOrThrow), () => {
-        it("should throw when there's no default export", () => {
-            const { sourceFile } = getInfoFromText("");
-            expect(() => sourceFile.getDefaultExportSymbolOrThrow()).to.throw();
-        });
-
-        it("should return the default export symbol when one exists", () => {
-            const { sourceFile } = getInfoFromText("export default class Identifier {}");
-            expect(sourceFile.getDefaultExportSymbolOrThrow().getName()).to.equal("default");
-        });
-    });
-
-    describe(nameof<SourceFile>(n => n.removeDefaultExport), () => {
-        it("should do nothing when there's no default export", () => {
-            const { sourceFile } = getInfoFromText("");
-            sourceFile.removeDefaultExport();
-            expect(sourceFile.getFullText()).to.equal("");
-        });
-
-        it("should remove the default export symbol when one exists", () => {
-            const { sourceFile } = getInfoFromText("export default class Identifier {}");
-            sourceFile.removeDefaultExport();
-            expect(sourceFile.getFullText()).to.equal("class Identifier {}");
-        });
-
-        it("should remove the default export symbol when default exported on a separate statement", () => {
-            const { sourceFile } = getInfoFromText("namespace Identifier {}\nclass Identifier {}\nexport default Identifier;\n");
-            sourceFile.removeDefaultExport();
-            expect(sourceFile.getFullText()).to.equal("namespace Identifier {}\nclass Identifier {}\n");
-        });
-    });
-
     describe(nameof<SourceFile>(n => n.getLanguageVariant), () => {
         it("should return standard when in a ts file", () => {
             const { sourceFile } = getInfoFromText("");
@@ -1004,7 +527,7 @@ describe(nameof(SourceFile), () => {
     });
 
     describe(nameof<SourceFile>(n => n.set), () => {
-        function doTest(startingCode: string, structure: SourceFileSpecificStructure, expectedCode: string) {
+        function doTest(startingCode: string, structure: Partial<SourceFileStructure>, expectedCode: string) {
             const { sourceFile } = getInfoFromText(startingCode);
             sourceFile.set(structure);
             expect(sourceFile.getFullText()).to.equal(expectedCode);
@@ -1016,19 +539,19 @@ describe(nameof(SourceFile), () => {
         });
 
         it("should modify when changed", () => {
-            const structure: MakeRequired<SourceFileSpecificStructure> = {
+            const structure: MakeRequired<SourceFileStructure> = {
                 imports: [{ moduleSpecifier: "module" }],
-                exports: [{ moduleSpecifier: "export-module" }]
+                exports: [{ moduleSpecifier: "export-module" }],
+                bodyText: "console.log()",
+                classes: [{ name: "C" }],
+                enums: [{ name: "E" }],
+                functions: [{ name: "F" }],
+                interfaces: [{ name: "I" }],
+                namespaces: [{ name: "N" }],
+                typeAliases: [{ name: "T", type: "string" }]
             };
-            doTest("import t from 'test'; export { test };", structure, `import "module";\n\nexport * from "export-module";\n`);
-        });
-
-        it("should remove when specifying empty arrays", () => {
-            const structure: MakeRequired<SourceFileSpecificStructure> = {
-                imports: [],
-                exports: []
-            };
-            doTest("import t from 'test'; export { test };", structure, "");
+            doTest("", structure, `import "module";\n\nclass C {\n}\n\nenum E {\n}\n\nfunction F() {\n}\n\ninterface I {\n}\n\nnamespace N {\n}\n\ntype T = string;\n\n` +
+                `export * from "export-module";\n`);
         });
     });
